+++ src/utils/index.js
@@ -19,3 +19,4 @@ export * from './generateSettingsLabel';
 export * from './stringToWords';
 export * from './getDropdownOptions';
 export * from './fetchJsonShemas';
+export * from './withProfileWrapper';